database/requirements.txt
kafka-python
psycopg2-binary
langdetect
requests